fix(dex+core): M6 F-040..F-043 — four generic semantic laws from the dooz scatter-map frontier
F-040 java.util.Arrays.fill family (bridge_to_api): OpenJDK fill + rangeCheck
law (IAE from>to; AIOOBE from<0/to>len) for all primitives + Object, 2-arg and
4-arg; value stored VERBATIM (wide/ref tags round-trip); null array -> deferred
NPE. Root cause of the dooz Lh/u;.a spin: Kotlin ArraysKt.fill(metadata, 0,
size, 0x8080808080808080) compiles to Ljava/util/Arrays;.fill([JIIJ)V —
missing handler = fail-soft no-op = all-zero scatter-map metadata = the SWAR
probe can never observe an EMPTY control byte. Live evidence: dooz
DerivedSnapshotState dependency put spun 38,179 Object.equals calls (rc=124
timeout); after the law rc=1 at the next honest layer.

F-041 encoded_catch_handler negative-size law (BOTH decode sites: throw
dispatch + find_catch_handler_for_pc): negative sleb size means |size| typed
pairs PLUS catch-all (DEX spec) — the old -(size+1) dropped one pair and
shifted the stream, decoding fixture handler '7f 14 6a d9 03' (typed type 20
-> 0x6a + catch-all 0x1d9) as catch-all @0x14 (mid-instruction garbage jump).
Affects every real R8/ECJ typed+catch-all handler entry.

F-042 VALUE_LONG static-default law: parse_static_values truncated encoded
long defaults via (int32_t) cast and materialization always made INT32; now
the parser keeps 64 bits and J-descriptor statics materialize as INT64.
Proven: the dooz-shape EMPTY marker 0x8080808080808080 round-trips (was
0x80808080).

F-043 java.lang.Double/Float IEEE bit-conversion law (bridge_to_api):
doubleTo(Raw)LongBits / longBitsToDouble / floatTo(Raw)IntBits /
intBitsToFloat with NaN canonicalization for non-Raw variants (OpenJDK
Double.java/Float.java); was REC-MISS -> null -> 0.

Evidence chain: androguard ground-truth disassembly of dooz Lh/u (androidx
collection ScatterMap family), Lh/r;.c insert commit, Ln/a;.r fill helper,
LF/F (DerivedSnapshotState); [F040-DIAG] env-gated wide-arg probe; micro-proof
f040_arrays_fill 7/7 bands GREEN, 3-run byte-identical; dooz rc 124 -> 1 with
the dependency-table insert now committing.

+++ miniandroid/src/dex/dex_parser.cpp
@@ -587,7 +587,13 @@ bool DexParser::parse_static_values(const uint8_t* data, uint32_t offset, ClassI
 
         if (v.has_int) {           // BYTE/SHORT/CHAR/INT/LONG (law-typed)
             field.has_default_value = true;
-            field.default_int_value = (int32_t)v.int_val;
+            // F-042 (MASTER-6 §C): keep the FULL 64-bit value. VALUE_LONG
+            // static defaults (e.g. `static long MARKER = 0x8080808080808080L`
+            // — the ScatterMap EMPTY constant shape) were truncated to 32
+            // bits here, so every sget-wide read back a half-word and the
+            // wide value never survived the invoke boundary. The field's
+            // declared descriptor decides the width at materialization.
+            field.default_int_value = v.int_val;
         } else if (v.has_float_bits) {  // F-028f: VALUE_FLOAT / VALUE_DOUBLE
             field.has_default_value = true;
             field.default_value_has_float_bits = true;
